requirements.txt
Django==2.2.2
#pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
django-cors-headers==2.5.2
django-filter==2.1.0
django-multiupload==0.5.2
django-phone-field==1.7.0
django-redis==4.10.0
django-webpack-loader==0.6.0
django-websocket-redis==0.5.2
djangorestframework==3.9.4
enum34==1.1.6
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
jsonplus==0.8.0
lxml==4.3.0
redis==3.2.1
requests==2.21.0
requests-unixsocket==0.1.5
tqdm==4.32.1
urllib3==1.24.2
